Refactor content and sighting handling: replace `lastConfirmedSighting` with `lostAt`, update field precision, and adjust commands, models, and tests for consistency.

+++ src/main/java/com/breece/trackrejoice/geo/LatLng.java
@@ -1,3 +1,5 @@
 package com.breece.trackrejoice.geo;
 
-public record LatLng(double lat, double lng) {}
+import java.math.BigDecimal;
+
+public record LatLng(BigDecimal lat, BigDecimal lng) {}
